Made lightboxes pretty and added error checking.
This commit customizes the css used by the jquery-ui to suit my taste.
It removes the compilation of the css for jquery-ui widget from the
rails asset pipeline and we over-ride the styles we want along with our
own css styles. The light-boxes are now much prettier and some
error-checking is added on the client to make sure lusers don't try to
post stupid things to the server.