Replace calls to `set` with `setProperties`.
This commit goes through places in the app where consecutive `set` calls
were being made and coalesces them into a single call to
`setProperties`. This also fixes the condition where a `selectedItem` on
the controller was not cleaned up properly on route transition when no
action was made on the selected Item.