Fix to remove stale references to `content`.
There was a bug in the filtering logic which would lead to stale
references to old `content` being retained on the `originalContent`
field of the filterableContentMixin. This commit fixes the issue by
explicitly forcing the route to call initialize on the mixin.
TODO: Kill this hack and refactor this on the ember_refactor PR that is
pending.